☕ Опыт Integer -> Double

+++ src/main/java/com/example/StudyFriends/dto/SkillDto.java
@@ -13,7 +13,7 @@ public class SkillDto {
     private Long playerId;
     private String name;
     private Integer level;
-    private Integer expAmount;
+    private Double expAmount;
     private Boolean isActive;
     private Boolean usedInSessions;
